pyproject.toml
[project]
name = "photo-email-scheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "pyyaml>=6.0.2",
]
